requirements.txt
tensorflow-cpu == 2.3.0
pyqt5~=5.15.9
pillow~=9.5.0
opencv-python~=4.7.0.72
matplotlib~=3.5.3
numpy~=1.18.5